Add support for moving room shares
When a recipient modifies a room share the share should be modified only
for that recipient and be left untouched for the other users in the
room. Therefore, when a recipient modifies a share a copy of the
original share is added to keep the share status (permissions and file
target) from the point of view of the recipient; when the share is later
accessed by the recipient RoomShareProvider returns it as seen by that
recipient.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>