Add support for creating rooms associated to arbitrary objects
Associating a room to another object is optional; by default a room is
not associated to any object.

The association is just a "flag" on the room; any special behaviour for
an association must be implemented as needed outside the room (for
example, by responding to events emitted by the room).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>