Add switch for enabling or disabling sending messages to Nextcloud log
This abuses the current signaling settings to be able to set whether the
messages should be sent to the Nextcloud log or not. As all this is
meant to be used in temporal patches just for testing this prevents
failures and retries if the patch is reverted, so there is no longer a
debug endpoint in the server, but the browsers still use a cached
version of the JavaScript files.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>