Hide videos and screens in the initial page
The "#videos" and "#screens" elements (which form the call view) were
visible in the initial page of the main Talk UI, so the local video view
was explicitly hidden to ensure that it was not shown. However, the
local video view should be shown whenever the call view is shown, and
the call view is the one that should be hidden in the initial page
instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>