Handle names in CallParticipantModel
CallParticipantModel only triggers change events when the model
attributes actually change, so the code to update the VideoView only
when the name has changed can be dropped.

Although before the "peer.nick" was got by listening to "videoAdded"
that value is now got by listening to "peerStreamAdded"; the former is
just a specialized event that is always triggered after the later.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>