Add explicit handling of a dark background in the CallView
The CallView background turns black when there are other participants in
a call or a screen ir shared. Some elements, like the sidebar trigger,
appear on the CallView, even if they are not children of it. When that
happens their style need to be adjusted, as their standard style is
designed for a white background.

Until now this was done based on the "incall" and "screensharing" CSS
classes set in the "#app-content" element, but now an explicit handling
independent of those classes has been added.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>